Export part of Header component into a container
