+++ src/reducers/index.js
@@ -1,7 +1,7 @@
-import { combineReducers } from 'redux';
-import { userInfo, loggedIn } from './facebook';
+import { combineReducers } from 'redux'
+import { userInfo, loggedIn } from './facebook'
 
 export default combineReducers({
     loggedIn,
     userInfo,
-});
+})
